requirements.txt
nltk==3.8.1
pyyaml
moviepy
pillow